Applied ncurses as the layout manager for console output.
Currently the screen is divided into three windows showing Messages, connection clients and scores. Scores and Client content occupies 1/4 and 1/4 respectively while the Message occupies 1/2 of the screen. Each panel is formed with a window(*_window) with a box, and a new window showing actual content(*_content) is created within the window(box) for content display. Scrolling is enabled on the message panel. wprintw() will be used instead of printf() in the system from now onwards catering for the new change.

	modified:   ../Makefile
	modified:   buzzer.c
	modified:   include/layout.h
	modified:   layout.c
	modified:   link_list.c
	modified:   non_blocking_socket.c
	modified:   quiz_server.c
	modified:   score.c
	modified:   score_db.c
	modified:   server_cb.c
	modified:   utilities.c