Refactoring and add doxygen documentation.
Added IPv6 support to the listening port 9000 by default.
Modify the linked list to insert into the head.
Info now contains the sockaddr_storage to store the copy of the
socket, instead of the char* host and port.
Added documentation to link_list.h and generated the documentation using
doxygen.